# Conflicts:
#	api/version_test.go
#	go.mod
#	go.sum

+++ Jenkinsfile
@@ -121,7 +121,7 @@ pipeline {
           revlist=`git rev-list --tags --max-count=1`
           rc=$?
           set -e
-          if [ 0 -eq $rc ]; then
+          if [ 0 -eq $rc -a x"$revlist" != x ]; then
             tag=`git describe --tags $revlist`
 
             major=`echo $tag | awk -F '.' '{ print $1 }'`
@@ -162,7 +162,7 @@ pipeline {
           revlist=`git rev-list --tags --max-count=1`
           rc=$?
           set -e
-          if [ 0 -eq $rc ]; then
+          if [ 0 -eq $rc -a x"$revlist" != x ]; then
             tag=`git describe --tags $revlist`
 
             major=`echo $tag | awk -F '.' '{ print $1 }'`
@@ -195,7 +195,7 @@ pipeline {
           revlist=`git rev-list --tags --max-count=1`
           rc=$?
           set -e
-          if [ 0 -eq $rc ]; then
+          if [ 0 -eq $rc -a x"$revlist" != x ]; then
             tag=`git describe --tags $revlist`
 
             major=`echo $tag | awk -F '.' '{ print $1 }'`
@@ -225,10 +225,15 @@ pipeline {
       }
       steps {
         sh(returnStdout: true, script: '''
+          set +e
           revlist=`git rev-list --tags --max-count=1`
-          tag=`git describe --tags $revlist`
-          git reset --hard
-          git checkout $tag
+          rc=$?
+          set -e
+          if [ 0 -eq $rc -a x"$revlist" != x ]; then
+            tag=`git describe --tags $revlist`
+            git reset --hard
+            git checkout $tag
+          fi
         '''.stripIndent())
         sh 'make verify-build'
         sh 'DEVELOPMENT=other DOCKER_REGISTRY=$DOCKER_REGISTRY make generate-docker-images'
@@ -256,15 +261,20 @@ pipeline {
       }
       steps {
         sh(returnStdout: false, script: '''
-          revlist=`git rev-list --tags --max-count=1`
-          tag=`git describe --tags $revlist`
-
           set +e
-          docker images | grep appuser-middleware | grep $tag
+          revlist=`git rev-list --tags --max-count=1`
           rc=$?
           set -e
-          if [ 0 -eq $rc ]; then
-            TAG=$tag DOCKER_REGISTRY=$DOCKER_REGISTRY make release-docker-images
+
+          if [ 0 -eq $rc -a x"$revlist" != x ]; then
+            tag=`git describe --tags $revlist`
+            set +e
+            docker images | grep appuser-middleware | grep $tag
+            rc=$?
+            set -e
+            if [ 0 -eq $rc ]; then
+              TAG=$tag DOCKER_REGISTRY=$DOCKER_REGISTRY make release-docker-images
+            fi
           fi
         '''.stripIndent())
       }
@@ -276,22 +286,28 @@ pipeline {
       }
       steps {
         sh(returnStdout: false, script: '''
-          revlist=`git rev-list --tags --max-count=1`
-          tag=`git describe --tags $revlist`
-
-          major=`echo $tag | awk -F '.' '{ print $1 }'`
-          minor=`echo $tag | awk -F '.' '{ print $2 }'`
-          patch=`echo $tag | awk -F '.' '{ print $3 }'`
-
-          patch=$(( $patch - $patch % 2 ))
-          tag=$major.$minor.$patch
-
           set +e
-          docker images | grep appuser-middleware | grep $tag
+          revlist=`git rev-list --tags --max-count=1`
           rc=$?
           set -e
-          if [ 0 -eq $rc ]; then
-            TAG=$tag DOCKER_REGISTRY=$DOCKER_REGISTRY make release-docker-images
+
+          if [ 0 -eq $rc -a x"$revlist" != x ]; then
+            tag=`git describe --tags $revlist`
+
+            major=`echo $tag | awk -F '.' '{ print $1 }'`
+            minor=`echo $tag | awk -F '.' '{ print $2 }'`
+            patch=`echo $tag | awk -F '.' '{ print $3 }'`
+
+            patch=$(( $patch - $patch % 2 ))
+            tag=$major.$minor.$patch
+
+            set +e
+            docker images | grep appuser-middleware | grep $tag
+            rc=$?
+            set -e
+            if [ 0 -eq $rc ]; then
+              TAG=$tag DOCKER_REGISTRY=$DOCKER_REGISTRY make release-docker-images
+            fi
           fi
         '''.stripIndent())
       }
